requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.8.0=py36_0
astor=0.8.0=py36_0
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
cairo=1.14.12=h8948797_3
certifi=2019.11.28=py36_1
cudatoolkit=9.2=0
cudnn=7.6.0=cuda9.2_0
cupti=9.2.148=0
cycler=0.10.0=py36_0
cython=0.29.13=py36he6710b0_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
gast=0.2.2=py36_0
glib=2.56.2=hd408876_0
gmp=6.1.2=h6c8ec71_1
google-pasta=0.1.7=py_0
graphite2=1.3.13=h23475e2_0
graphviz=2.40.1=h21bd128_2
grpcio=1.16.1=py36hf8bcb03_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=py36h7918eee_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h9c2bf20_1
intel-openmp=2019.4=243
ipykernel=5.1.2=py36h39e3cac_0
ipython=7.8.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36_0
ipywidgets=7.5.1=py_0
jedi=0.15.1=py36_0
jinja2=2.10.1=py36_0
joblib=0.13.2=py36_0
jpeg=9b=h024ee3a_2
json5=0.8.5=py_0
jsonschema=3.0.2=py36_0
jupyter=1.0.0=py36_7
jupyter_client=5.3.3=py_0
jupyter_console=6.0.0=py36_0
jupyter_core=4.5.0=py_0
jupyterlab=1.2.4=py_0
jupyterlab_server=1.0.6=py_0
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py36_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py36he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libmr=0.1.9=pypi_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.9.2=hd408876_0
libsodium=1.0.16=h1bed415_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
line_profiler=2.1.2=py36h14c3975_0
llvmlite=0.29.0=py36hd408876_0
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36h7b6447c_0
matplotlib=3.1.1=py36h5429711_0
mistune=0.8.4=py36h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
nbconvert=5.6.0=py36_1
nbformat=4.4.0=py36_0
ncurses=6.1=he6710b0_1
notebook=6.0.1=py36_0
numba=0.45.1=py36h962f231_0
numpy=1.17.2=py36haad9e8e_0
numpy-base=1.17.2=py36hde5b4d6_0
openssl=1.1.1f=h7b6447c_0
pandas=0.25.1=py36he6710b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
pango=1.42.4=h049681c_0
parso=0.5.1=py_0
patsy=0.5.1=py36_0
pcre=8.43=he6710b0_0
pexpect=4.7.0=py36_0
pickleshare=0.7.5=py36_0
pip=19.2.3=py36_0
pixman=0.38.0=h7b6447c_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.9.2=py36he6710b0_0
ptyprocess=0.6.0=py36_0
pydot=1.4.1=py36_0
pygments=2.4.2=py_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py36h05f1152_2
pyrsistent=0.14.11=py36h7b6447c_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py36_0
pytz=2019.2=py_0
pyyaml=5.1.2=py36h7b6447c_0
pyzmq=18.1.0=py36he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.5.5=py_0
readline=7.0=h7b6447c_5
scikit-commpy=0.4.0=pypi_0
scikit-learn=0.21.2=py36hd81dba3_0
scipy=1.3.1=py36h7c811a0_0
seaborn=0.9.0=py36_0
send2trash=1.5.0=py36_0
setuptools=41.2.0=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.29.0=h7b6447c_0
statsmodels=0.10.1=py36hdd07704_0
tbb=2019.8=hfd86e86_0
tensorboard=1.14.0=py36hf484d3e_0
tensorflow=1.14.0=gpu_py36hfc5689a_0
tensorflow-base=1.14.0=gpu_py36h611c6d2_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
terminado=0.8.2=py36_0
testpath=0.4.2=py36_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h7b6447c_0
traitlets=4.3.2=py36_0
wcwidth=0.1.7=py36_0
webencodings=0.5.1=py36_1
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
widgetsnbextension=3.5.1=py36_0
wrapt=1.11.2=py36h7b6447c_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zeromq=4.3.1=he6710b0_3
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
